更新 FIFO 缓冲区函数签名，将 size 和 length 参数类型从 size_t 修改为 uint32_t，以提高一致性和兼容性

+++ src/lib/buffer/fifo_buffer.h
@@ -9,8 +9,8 @@ typedef void (*fifo_buffer_mutex_unlock_t)(void *mutex);
 typedef struct {
     uint8_t *buffer;
     size_t size;
-    size_t head;
-    size_t tail;
+    size_t head; // write index
+    size_t tail; // read index
 #ifdef FIFO_BUFFER_USING_MUTEX
     void *mutex;
     fifo_buffer_mutex_lock_t lock;
@@ -19,16 +19,16 @@ typedef struct {
 } fifo_buffer_t;
 
 #ifdef FIFO_BUFFER_USING_MUTEX
-int fifo_buffer_init(fifo_buffer_t *fifo, uint8_t *buffer, size_t size, void *mutex, fifo_buffer_mutex_lock_t lock, fifo_buffer_mutex_unlock_t unlock);
+int fifo_buffer_init(fifo_buffer_t *fifo, uint8_t *buffer, uint32_t size, void *mutex, fifo_buffer_mutex_lock_t lock, fifo_buffer_mutex_unlock_t unlock);
 #else
-int fifo_buffer_init(fifo_buffer_t *fifo, uint8_t *buffer, size_t size);
+int fifo_buffer_init(fifo_buffer_t *fifo, uint8_t *buffer, uint32_t size);
 #endif
 
-int fifo_buffer_write(fifo_buffer_t *fifo, const uint8_t *data, size_t length);
+int fifo_buffer_write(fifo_buffer_t *fifo, const uint8_t *data, uint32_t length);
 
-int fifo_buffer_read(fifo_buffer_t *fifo, uint8_t *data, size_t length);
+int fifo_buffer_read(fifo_buffer_t *fifo, uint8_t *data, uint32_t length);
 
-int fifo_buffer_peek(fifo_buffer_t *fifo, uint8_t *data, size_t length, size_t offset);
+int fifo_buffer_peek(fifo_buffer_t *fifo, uint8_t *data, uint32_t length, uint32_t offset);
 
 int fifo_buffer_get_left(fifo_buffer_t *fifo);
